+++ src/vs/workbench/contrib/accessibility/browser/accessibleView.ts
@@ -23,7 +23,8 @@ import { IContextViewDelegate, IContextViewService } from 'vs/platform/contextvi
 import { IInstantiationService, createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { alert } from 'vs/base/browser/ui/aria/aria';
-import { IEditorOptions } from 'vs/editor/common/config/editorOptions';
+import { getSimpleEditorOptions } from 'vs/workbench/contrib/codeEditor/browser/simpleEditorOptions';
+import { SelectionClipboardContributionID } from 'vs/workbench/contrib/codeEditor/browser/selectionClipboard';
 
 const enum DEFAULT {
 	WIDTH = 800,
@@ -79,7 +80,7 @@ class AccessibleView extends Disposable {
 		this._editorContainer = document.createElement('div');
 		this._editorContainer.classList.add('accessible-view');
 		const codeEditorWidgetOptions: ICodeEditorWidgetOptions = {
-			contributions: EditorExtensionsRegistry.getSomeEditorContributions([LinkDetector.ID, 'editor.contrib.selectionClipboard', 'editor.contrib.selectionAnchorController'])
+			contributions: EditorExtensionsRegistry.getSomeEditorContributions([LinkDetector.ID, SelectionClipboardContributionID, 'editor.contrib.selectionAnchorController'])
 		};
 		const editorOptions: IEditorConstructionOptions = {
 			...getSimpleEditorOptions(this._configurationService),
@@ -225,35 +226,4 @@ export class AccessibleViewService extends Disposable implements IAccessibleView
 		}
 		this._accessibleView.show(provider);
 	}
-}
-function getSimpleEditorOptions(configurationService: IConfigurationService): IEditorOptions {
-	return {
-		wordWrap: 'on',
-		overviewRulerLanes: 0,
-		glyphMargin: false,
-		lineNumbers: 'off',
-		folding: false,
-		selectOnLineNumbers: false,
-		hideCursorInOverviewRuler: true,
-		selectionHighlight: false,
-		scrollbar: {
-			horizontal: 'hidden'
-		},
-		lineDecorationsWidth: 0,
-		overviewRulerBorder: false,
-		scrollBeyondLastLine: false,
-		renderLineHighlight: 'none',
-		fixedOverflowWidgets: true,
-		acceptSuggestionOnEnter: 'smart',
-		dragAndDrop: false,
-		revealHorizontalRightPadding: 5,
-		minimap: {
-			enabled: false
-		},
-		guides: {
-			indentation: false
-		},
-		accessibilitySupport: configurationService.getValue<'auto' | 'off' | 'on'>('editor.accessibilitySupport'),
-		cursorBlinking: configurationService.getValue<'blink' | 'smooth' | 'phase' | 'expand' | 'solid'>('editor.cursorBlinking')
-	};
 }
